refactor: merge check Auth middlewares into one genaric middleware
