+++ server/utils/index.ts
@@ -2,3 +2,4 @@ export { signupSchema, dishSchema, loginSchema } from './validations/joi';
 export { default as CustomError } from './helper/customError';
 export * from './helper';
 export * from './types';
+export * from './enum';
