+++ server/utils/types/index.ts
@@ -6,4 +6,4 @@ export { default as CartAttributes } from './cart';
 export { default as OrderAttributes } from './order';
 export { default as staffAttribute } from './staffAttributes';
 export { default as Payload } from './payload';
-export { default as CustomRequest } from './payload';
+export { default as CustomRequest } from './authInterface';
